refactor(app): Replaced `DrawerLayout` with `BottomNavigationView`. Restart app replaced with navigation. Order handling improved
Instead of restarting app on sign out, navigation to RegistrationFragment is done. UID-based cart handling done to ensure that multiple users on the same device do not lose their cart.
While placing an order, if the cart contains items spanning across both the days, it is split into two orders.